server-side/requirements.txt
Django == 4.2.6
djangorestframework == 3.14.0
django-cors-headers == 4.3.0
google-api-python-client == 2.103.0
google-auth-httplib2 == 0.1.1
google-auth-oauthlib == 1.1.0